returned old settings

+++ Jenkinsfile.groovy
@@ -1,31 +1,24 @@
 def eurekaServer = new Structure(
         "eureka-server",
-        "structural/eureka-server/build/Dockerfile",
-        "127.0.0.1:8761:8761")
+        "structural/eureka-server/build/Dockerfile")
 def authService = new Structure(
         "auth-service",
-        "structural/auth-service/build/Dockerfile",
-        "127.0.0.1:8000:8000")
+        "structural/auth-service/build/Dockerfile")
 def zuulGateway = new Structure(
         "zuul-gateway",
-        "structural/zuul-gateway/build/Dockerfile",
-        "8001:8001")
+        "structural/zuul-gateway/build/Dockerfile")
 def documentationService = new Structure(
         "documentation-service",
-        "structural/documentation-service/build/Dockerfile",
-        "127.0.0.1:8002:8002")
+        "structural/documentation-service/build/Dockerfile")
 def userService = new Structure(
         "user-service",
-        "business/user-service/build/Dockerfile",
-        "127.0.0.1:8010:8010")
+        "business/user-service/build/Dockerfile")
 def selfSovereignIdentityService = new Structure(
         "self-sovereign-identity-service",
-        "business/self-sovereign-identity-service/build/Dockerfile",
-        "127.0.0.1:8011:8011")
+        "business/self-sovereign-identity-service/build/Dockerfile")
 def governmentService = new Structure(
         "government-service",
-        "business/fake-government-service/build/Dockerfile",
-        "127.0.0.1:8012:8012")
+        "business/fake-government-service/build/Dockerfile")
 
 def modules = [
         eurekaServer,
@@ -69,38 +62,35 @@ pipeline {
 
                     for (int i = 0; i < modules.size(); i++) {
                         serviceName = modules[i].serviceName
-                        // if (gitChanges.contains(serviceName)) {
+                        if (gitChanges.contains(serviceName)) {
 
-                        int checkIfContainerExists = sh(
-                                script: 'docker ps -f name=' + serviceName + ' | grep -w ' + serviceName + ' | wc -l',
-                                returnStdout: true
-                        ).trim()
+                            int checkIfContainerExists = sh(
+                                    script: 'docker ps -f name=' + serviceName + ' | grep -w ' + serviceName + ' | wc -l',
+                                    returnStdout: true
+                            ).trim()
 
-                        if (checkIfContainerExists != 0) {
-                            removeContainerCommand = 'docker rm \$(docker stop ' + serviceName + ')'
-                            println removeContainerCommand
-                            sh removeContainerCommand
+                            if (checkIfContainerExists != 0) {
+                                removeContainerCommand = 'docker rm \$(docker stop ' + serviceName + ')'
+                                println removeContainerCommand
+                                sh removeContainerCommand
 
-                            removeImageCommand = 'docker rmi medical-ledger/' + serviceName
-                            println removeImageCommand
-                            sh removeImageCommand
+                                removeImageCommand = 'docker rmi medical-ledger/' + serviceName
+                                println removeImageCommand
+                                sh removeImageCommand
+                            }
+
+                            buildCommand = 'docker build -t medical-ledger/' + serviceName + ' -f ' + modules[i].dockerPath + ' .'
+                            println buildCommand
+                            sh buildCommand
+
+                            String volume = " -v /var/ledger:/var/ledger"
+                            if (serviceName.equals("government-service"))
+                                volume = " -v /var/government:/var/government"
+
+                            runCommand = 'docker run -d --name ' + serviceName + volume + ' --net=host medical-ledger/' + serviceName
+                            println runCommand
+                            sh runCommand
                         }
-
-                        buildCommand = 'docker build -t medical-ledger/' + serviceName + ' -f ' + modules[i].dockerPath + ' .'
-                        println buildCommand
-                        sh buildCommand
-
-                        String volume = " -v /var/ledger:/var/ledger"
-                        if (serviceName == "government-service")
-                            volume = " -v /var/government:/var/government"
-
-                        String port = modules[i].port
-                        println port
-
-                        runCommand = 'docker run -d --name ' + serviceName + volume + ' -p ' + port + ' medical-ledger/' + serviceName
-                        println runCommand
-                        sh runCommand
-                        // }
                     }
                 }
             }
@@ -111,11 +101,9 @@ pipeline {
 class Structure {
     String serviceName
     String dockerPath
-    String port
 
-    Structure(String serviceName, String dockerPath, String port) {
+    Structure(String serviceName, String dockerPath) {
         this.dockerPath = dockerPath
         this.serviceName = serviceName
-        this.port = port
     }
 }
